+++ dwds/lib/asset_handler.dart
@@ -8,9 +8,20 @@ import 'package:shelf/shelf.dart';
 import 'package:shelf_proxy/shelf_proxy.dart';
 
 /// A handler for a build target's assets.
+abstract class AssetHandler {
+  Handler get handler;
+
+  /// Returns the asset from a relative [path].
+  ///
+  /// For example the path `main.dart` should return the raw text value of that
+  /// corresponding file.
+  Future<Response> getRelativeAsset(String path);
+}
+
+/// A handler for a build target's assets.
 ///
-/// Proxies requests to the build daemon asset server.
-class AssetHandler {
+/// Proxies requests to the build runner's asset server.
+class BuildRunnerAssetHandler implements AssetHandler {
   final int _assetServerPort;
   final String _target;
   final int _applicationPort;
@@ -18,16 +29,14 @@ class AssetHandler {
 
   Handler _handler;
 
-  AssetHandler(this._assetServerPort, this._target, this._applicationHost,
-      this._applicationPort);
+  BuildRunnerAssetHandler(this._assetServerPort, this._target,
+      this._applicationHost, this._applicationPort);
 
+  @override
   Handler get handler =>
       _handler ??= proxyHandler('http://localhost:$_assetServerPort/$_target/');
 
-  /// Returns the asset from a relative [path].
-  ///
-  /// For example the path `main.dart` should return the raw text value of that
-  /// corresponding file.
+  @override
   Future<Response> getRelativeAsset(String path) async => handler(Request(
       'GET', Uri.parse('http://$_applicationHost:$_applicationPort/$path')));
 }