+++ webdev/CHANGELOG.md
@@ -1,8 +1,8 @@
 ## 2.5.2-dev
 
 - Update SDK dependency to minimum of 2.5.0.
-
 - Support absolute file: URIs for libraries.
+- Depend on the latest `package:dwds`.
 
 ## 2.5.1